Add helper for blob URLs in tests
